+++ liveviz/src/app/App.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import AppUI from '../ui/AppUI'
+import AppUI from './AppUI'
 import { setServerRoot, setConnectionStatus } from '../actions/GlobalActions'
 
 const mapDispatchToProps = (dispatch) => ({
