+++ liveviz/src/app/trackerApp/TrackerAppUI.js
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Grid, Col } from 'react-bootstrap'
-import TitleBarUI from './TitleBarUI'
-import GraphSectionUI from './GraphSectionUI'
-import Params from '../logic/Params'
+import TitleBarUI from '../../ui/TitleBarUI'
+import GraphSectionUI from '../../ui/GraphSectionUI'
+import Params from './params/logic/Params'
 
-import '../styles/params.css'
-import '../styles/graph.css'
+import '../../styles/params.css'
+import '../../styles/graph.css'
 
 const GraphAndParamsUI = ({ multipleMapView, singleMapView, view, url, status, displayParams, running, updateStatus, setViewMode, toggleRunning, toggleParams }) =>
 <Grid fluid={true} id="graph-params" className={"" + (displayParams? "paramsDisplayed" : "paramsHidden") }>
